Move ICoreModelCapability from core to core.api bundle.
