+++ clientprovider/src/test/java/org/mqnaas/clientprovider/impl/apiclient/APIProviderFactoryTest.java
@@ -7,8 +7,8 @@ import org.junit.Test;
 import org.mqnaas.bundletree.IBundleGuard;
 import org.mqnaas.clientprovider.impl.AbstractProviderFactory;
 import org.mqnaas.core.api.Endpoint;
+import org.mqnaas.core.api.ICoreModelCapability;
 import org.mqnaas.core.api.IRootResource;
-import org.mqnaas.core.impl.ICoreModelCapability;
 import org.mqnaas.test.helpers.ReflectionTestHelper;
 import org.mqnaas.test.helpers.capability.ArtificialBundleGuard;
 import org.mqnaas.test.helpers.capability.TestCapabilitiesFactory;
