+++ core.api/src/main/java/org/mqnaas/core/api/ICoreModelCapability.java
@@ -1,8 +1,5 @@
-package org.mqnaas.core.impl;
+package org.mqnaas.core.api;
 
-import org.mqnaas.core.api.ICapability;
-import org.mqnaas.core.api.IResource;
-import org.mqnaas.core.api.IRootResource;
 
 /**
  * This {@link ICapability} groups all the operations that can be requested to core implementation.
