+++ core/src/main/java/org/mqnaas/core/impl/BindingManagement.java
@@ -16,6 +16,7 @@ import org.mqnaas.core.api.Endpoint;
 import org.mqnaas.core.api.IApplication;
 import org.mqnaas.core.api.IBindingDecider;
 import org.mqnaas.core.api.ICapability;
+import org.mqnaas.core.api.ICoreModelCapability;
 import org.mqnaas.core.api.IExecutionService;
 import org.mqnaas.core.api.IObservationService;
 import org.mqnaas.core.api.IResource;
