+++ test-helpers/pom.xml
@@ -26,10 +26,6 @@
 		</dependency>
 		<dependency>
 			<groupId>org.mqnaas</groupId>
-			<artifactId>core</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.mqnaas</groupId>
 			<artifactId>bundletree</artifactId>
 		</dependency>
 		<dependency>
